backend/requirements.txt
# Core FastAPI stack
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
pydantic==2.5.0

# Managed database clients
supabase==2.3.0
neo4j==5.14.0
pinecone-client==2.2.4
redis==5.0.5

# AI/ML libraries
google-generativeai==0.3.0
openai==1.3.0
sentence-transformers==2.2.2

# HTTP and data processing
httpx==0.25.2
xmltodict==0.13.0

# PDF and document processing
PyMuPDF==1.23.0
pdf2image==1.16.3
Pillow==10.1.0

# Background job processing
upstash-redis==0.15.0

# Utility libraries
asyncio-throttle==1.0.2
python-multipart==0.0.6